requirements.txt
mcp[cli]
langfuse==2.60.8
fastapi
uvicorn
elasticsearch==8.17.1
autogen-agentchat
autogen-ext[azure]
duckduckgo_search==6.4.2
wikipedia
aiohttp
sentence_transformers
neo4j 
langchain-neo4j
langchain-openai
ragas
llama_cloud_services
llama-index-embeddings-huggingface
json_schema_to_pydantic
langchain_neo4j
python-multipart